app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.9.2